move tests for output formatting to the relevant module
